+++ src/components/trip-info.tsx
@@ -157,6 +157,8 @@ export function TripInfo({
       )}
       <RouteBadge
         route={route}
+        dataLang={dataLang}
+        agencyLangs={agencyLang}
         infoLevel={infoLevel}
         size={size === 'sm' ? 'sm' : undefined}
         disableVerbose={true}